requirements.txt
openai
python-dotenv
pandas
networkx
sentence-transformers
openpyxl
streamlit
